+++ Source/CDR.DataRecipient.Repository.SQL/Migrations/20220228033632_DataHolder_Update.cs
@@ -12,9 +12,8 @@ namespace CDR.DataRecipient.Repository.SQL.Migrations
             migrationBuilder.AddColumn<DateTime>(
                 name: "LastUpdated",
                 table: "DataHolderBrand",
-                type: "datetime2",
-                nullable: false,
-                defaultValue: new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified));
+                type: "datetime",
+                nullable: true);
         }
 
         protected override void Down(MigrationBuilder migrationBuilder)