+++ Source/CDR.DataRecipient.Repository.SQL/Migrations/20220311045322_Registration_DHBrandId_Update.cs
@@ -1,0 +1,27 @@
+﻿using System;
+using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace CDR.DataRecipient.Repository.SQL.Migrations
+{
+    public partial class Registration_DHBrandId_Update : Migration
+    {
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.AddColumn<Guid>(
+                name: "DataHolderBrandId",
+                table: "Registration",
+                type: "uniqueidentifier",
+                nullable: true,
+                defaultValue: null);
+        }
+
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropColumn(
+                name: "DataHolderBrandId",
+                table: "Registration");
+        }
+    }
+}