+++ Source/CDR.DataRecipient.Repository.SQL/Migrations/20220620050738_UpdateDcrMsg_Infosec_Created.cs
@@ -1,0 +1,38 @@
+﻿using System;
+using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace CDR.DataRecipient.Repository.SQL.Migrations
+{
+    public partial class UpdateDcrMsg_Infosec_Created : Migration
+    {
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.AddColumn<DateTime>(
+                name: "Created",
+                table: "DcrMessage",
+                type: "datetime2",
+                nullable: false,
+                defaultValue: new DateTime(1, 1, 1, 0, 0, 0, 0, DateTimeKind.Unspecified));
+
+            migrationBuilder.AddColumn<string>(
+                name: "InfosecBaseUri",
+                table: "DcrMessage",
+                type: "nvarchar(500)",
+                maxLength: 500,
+                nullable: true);
+        }
+
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropColumn(
+                name: "Created",
+                table: "DcrMessage");
+
+            migrationBuilder.DropColumn(
+                name: "InfosecBaseUri",
+                table: "DcrMessage");
+        }
+    }
+}