+++ Source/CDR.DataRecipient.Web/Controllers/ParController.cs
@@ -3,7 +3,6 @@ using CDR.DataRecipient.Repository;
 using CDR.DataRecipient.SDK.Models;
 using CDR.DataRecipient.SDK.Services.DataHolder;
 using CDR.DataRecipient.Web.Common;
-using CDR.DataRecipient.Web.Configuration;
 using CDR.DataRecipient.Web.Extensions;
 using CDR.DataRecipient.Web.Filters;
 using CDR.DataRecipient.Web.Models;
@@ -64,70 +63,86 @@ namespace CDR.DataRecipient.Web.Controllers
         {
             if (!string.IsNullOrEmpty(model.ClientId))
             {
-                var reg = await _registrationsRepository.GetRegistration(model.ClientId);
-                var dhConfig = await _dataHolderDiscoveryCache.GetOidcDiscoveryByBrandId(reg.DataHolderBrandId);
-                var sp = _config.GetSoftwareProductConfig();
-                var infosecBaseUri = await GetInfoSecBaseUri(reg.DataHolderBrandId);
-                var stateKey = Guid.NewGuid().ToString();
-                var nonce = Guid.NewGuid().ToString();
-                var redirectUri = model.RedirectUris;
+                try {
+                    var reg = await _registrationsRepository.GetRegistration(model.ClientId);
+                    var dhConfig = await _dataHolderDiscoveryCache.GetOidcDiscoveryByBrandId(reg.DataHolderBrandId);
+                    var sp = _config.GetSoftwareProductConfig();
 
-                var authState = new AuthorisationState()
-                {
-                    StateKey = stateKey,
-                    ClientId = model.ClientId,
-                    SharingDuration = model.SharingDuration,
-                    Scope = model.Scope,
-                    DataHolderBrandId = reg.DataHolderBrandId,
-                    DataHolderInfosecBaseUri = infosecBaseUri,
-                    RedirectUri = redirectUri
-                };
+                    var infosecBaseUri = await GetInfoSecBaseUri(reg.DataHolderBrandId);
+                    if (string.IsNullOrEmpty(infosecBaseUri))
+                        throw new CustomException();
 
-                if (model.UsePkce)
-                {
-                    authState.Pkce = _dhInfoSecService.CreatePkceData();
-                }
+                    var stateKey = Guid.NewGuid().ToString();
+                    var nonce = Guid.NewGuid().ToString();
+                    var redirectUri = model.RedirectUris;
 
-                await _cache.SetAsync(stateKey, authState, DateTimeOffset.UtcNow.AddMinutes(60));
+                    var authState = new AuthorisationState()
+                    {
+                        StateKey = stateKey,
+                        ClientId = model.ClientId,
+                        SharingDuration = model.SharingDuration,
+                        Scope = model.Scope,
+                        DataHolderBrandId = reg.DataHolderBrandId,
+                        DataHolderInfosecBaseUri = infosecBaseUri,
+                        RedirectUri = redirectUri,
+                        UserId = this.HttpContext.User.GetUserId()
+                    };
 
-                // Build the authentication request JWT.
-                var authRequest = _dhInfoSecService.BuildAuthorisationRequestJwt(
-                    dhConfig.Issuer,
-                    model.ClientId,
-                    redirectUri,
-                    model.Scope,
-                    stateKey,
-                    nonce,
-                    sp.SigningCertificate.X509Certificate,
-                    model.SharingDuration,
-                    model.CdrArrangementId,
-                    "form_post",
-                    authState.Pkce);
+                    if (model.UsePkce)
+                        authState.Pkce = _dhInfoSecService.CreatePkceData();
 
-                var parResponse = await _dhInfoSecService.PushedAuthorisationRequest(
-                    dhConfig.PushedAuthorizationRequestEndpoint,
-                    sp.ClientCertificate.X509Certificate,
-                    sp.SigningCertificate.X509Certificate,
-                    model.ClientId,
-                    authRequest);
+                    await _cache.SetAsync(stateKey, authState, DateTimeOffset.UtcNow.AddMinutes(60));
 
-                model.StatusCode = parResponse.StatusCode;
-                model.Messages = parResponse.Message;
-
-                if (parResponse.IsSuccessful)
-                {
-                    model.PushedAuthorisation = parResponse.Data;
-
-                    // Build the Authorisation URL for the Data Holder passing in the request uri returned from the PAR response.
-                    model.AuthorisationUri = await _dhInfoSecService.BuildAuthorisationRequestUri(
-                        infosecBaseUri,
+                    // Build the authentication request JWT.
+                    var authRequest = _dhInfoSecService.BuildAuthorisationRequestJwt(
+                        dhConfig.Issuer,
                         model.ClientId,
+                        redirectUri,
+                        model.Scope,
+                        stateKey,
+                        nonce,
                         sp.SigningCertificate.X509Certificate,
-                        model.PushedAuthorisation.RequestUri);
+                        model.SharingDuration,
+                        model.CdrArrangementId,
+                        "form_post",
+                        authState.Pkce);
+
+                    var parResponse = await _dhInfoSecService.PushedAuthorisationRequest(
+                        dhConfig.PushedAuthorizationRequestEndpoint,
+                        sp.ClientCertificate.X509Certificate,
+                        sp.SigningCertificate.X509Certificate,
+                        model.ClientId,
+                        authRequest);
+
+                    model.StatusCode = parResponse.StatusCode;
+                    model.Messages = parResponse.Message;
+
+                    if (parResponse.IsSuccessful)
+                    {
+                        model.PushedAuthorisation = parResponse.Data;
+
+                        // Build the Authorisation URL for the Data Holder passing in the request uri returned from the PAR response.
+                        model.AuthorisationUri = await _dhInfoSecService.BuildAuthorisationRequestUri(
+                            infosecBaseUri,
+                            model.ClientId,
+                            sp.SigningCertificate.X509Certificate,
+                            model.PushedAuthorisation.RequestUri,
+                            model.Scope);
+                    }
+                    else
+                    {
+                        model.ErrorList = parResponse.Errors;
+                    }
                 }
-                else
+                catch (CustomException)
                 {
-                    model.ErrorList = parResponse.Errors;
+                    var msg = $"The Data Holder details do not exist in the repository for ClientId: {model.ClientId}";
+                    return View("Error", new ErrorViewModel { Message = msg });
+                }
+                catch (Exception ex)
+                {
+                    var msg = $"Unable to create the Pushed Authorisation Request (PAR) with ClientId: {model.ClientId} - {ex.Message}";
+                    return View("Error", new ErrorViewModel { Message = msg });
                 }
             }
 
@@ -179,7 +194,7 @@ namespace CDR.DataRecipient.Web.Controllers
             if (model.Registrations != null && model.Registrations.Any())
             {
                 model.RegistrationListItems = model.Registrations
-                    .Select(r => new SelectListItem($"DH Brand: {r.DataHolderBrandId} ({r.ClientId})", r.ClientId))
+                    .Select(r => new SelectListItem($"DH Brand: {r.BrandName} ({r.DataHolderBrandId}) - ({r.ClientId})", r.ClientId))
                     .ToList();
             }
 
@@ -189,6 +204,9 @@ namespace CDR.DataRecipient.Web.Controllers
         private async Task<string> GetInfoSecBaseUri(string dataHolderBrandId)
         {
             var dh = await _dhRepository.GetDataHolderBrand(dataHolderBrandId);
+            if (dh == null)
+                return null;
+
             return dh.EndpointDetail.InfoSecBaseUri;
         }
     }