+++ Source/CDR.DataRecipient.Web/Controllers/UtilitiesController.cs
@@ -1,12 +1,16 @@
 ﻿using CDR.DataRecipient.SDK.Extensions;
 using CDR.DataRecipient.SDK.Register;
-using CDR.DataRecipient.Web.Configuration;
+using CDR.DataRecipient.SDK.Services.Register;
+using CDR.DataRecipient.Web.Caching;
+using CDR.DataRecipient.Web.Common;
+using CDR.DataRecipient.Web.Extensions;
 using CDR.DataRecipient.Web.Filters;
 using CDR.DataRecipient.Web.Models;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Configuration;
 using System;
+using System.Threading.Tasks;
 
 namespace CDR.DataRecipient.Web.Controllers
 {
@@ -14,43 +18,15 @@ namespace CDR.DataRecipient.Web.Controllers
     [Route("utilities")]
     public class UtilitiesController : Controller
     {
-        public const string DEFAULT_PRIVATE_KEY =
-            @"-----BEGIN PRIVATE KEY-----
-MIIEvwIBADANBgkqhkiG9w0BAQEFAASCBKkwggSlAgEAAoIBAQDqSkpDrvhkPp0K
-slpl7NqHgemvwPnGjg+fhOJsO1apavjaCGrZJ0nT2VoeKXiHW3727dgtst41CVur
-aLPK9C/+XyZZ8FGpvv5SAR9PgQUCWuIPmHzjGxuwrzmVrPumHssDySPXXO3G5SqN
-4ZuKtyeDa/oYzyfMxKTyfiD2c1LLf4bSdlw9g+phxshLA+IVf1E3ho8rKyi64nWT
-UghV0/NRVoBDA/kX1HtMTPWVaQPguPWfD6bHonF7l+KjO4xLzusPwD0Be2477num
-Zb+bjR+IMlEXtyxg5gvU3fYBRwgislbQph0iZa1o4yb6YACVsZfbPPfOdcSUpoSb
-SsGpl3HlAgMBAAECggEBAN+LQK6xzNjBEVA6epxT0PKIVzNVujzkIS4AOZYxS+/c
-XFKUw3Ys0XlsIlszEIh+GXd72s9HolMSeb/j5+CW+xAHM22PKiv/S2NtJLXUEbZ8
-hsOAqHB81f/QQO56HaYULXqQOT1ssoca825qu+Ev/mib5wYy6fOsEQI6rDLaOD5m
-vQ0CauaTjQdOZJ7CarMJeIz1HphwN+veW/YWuYNLmNsdn0B0/LFXMRunZ69wovfI
-4+fn2tjakqTv7FjdbK+ZXrQIyDQst/3i4f5ttgnfcgCa21aSEnC39+dV9VDyQXJc
-D4fo27c2dqzjnMcITl5LQ1oe1fjIjIQ+Y6OJX+pR7QECgYEA+VzvvZwn41PPTta+
-1PekS3Uv3OKfxW8Jk8HCxvJLJs0E7lwiJsl+bpv4IwS9S7mQ62XN8GpVEa1fooRa
-ZUj3zks6ISvPz8Gqq0WNTSCcRQckUBNqkVqVlrV4Obruz9Ymu1Grm1mUiRAYUoYQ
-3uaimUrCdmmdHM7R6MA4gpzhZcECgYEA8IanMht+NrzTm4Ljnv7D59F0dysTrLV/
-xgGlARSlsH2gi5F01MxCSDQV4FiXG8+LyWNns4lmmFuIdVbImT2tt44OQsn4WxFi
-17i7YMq56bigd0klVg76w5tdBVP1dJSHDBOBk6VpXUnV+qkkNXFljwjuN46BIRDO
-zuIe/CB0/SUCgYEAsM+R+grwSYMSml8wHoedShfEoUVbbj2mN8uKlVAVs2Rpm61e
-VcxHRpx23DWvFzNzq0WbOV3cBdW92tkn02tisjaq9/w9tJ0oq5p8b3Sw+UzwFYs+
-4+Or75mqrpx6WooJGob1PAjPhkQQSuteqP41yqW0rwuB6HxJYFRzfUFJnMECgYEA
-4qzT9xoH6YgtGJrS722DrP4td31GTnbCUxYLrigKOnk6iy3q3/0b3jYQA90Zk6EW
-SRYAjifIY3+n64V9CUYiaCFdeT4ka5bIAytak86aRRS1TZXGtQLq+tt5X+MPKO4t
-E4lyjXXPBZbnLRKoSCF1J8av0fXf5gyUCk76CnyAJG0CgYA38AmsG4/z83G0g8RG
-3i9Jzxw7QCVJfaibZ5hoSsK61wMHHFW4JrlO7uhIOd4bn0iV2Mqj1bD6Kg4Tfxc3
-//TigAqQ6M46sweOglFlFkKmBtAmjsfBpHMSOyIedCxCBYQESeW4BdqXMU+OqUnZ
-RIj2P/wtcJRnuztcszBJsnt7NQ==
------END PRIVATE KEY-----
-";
-
         private readonly IConfiguration _config;
+        private readonly ICacheManager _cacheManager;
 
         public UtilitiesController(
-            IConfiguration config)
+            IConfiguration config,
+            ICacheManager cacheManager)
         {
             _config = config;
+            _cacheManager = cacheManager;
         }
 
         [HttpGet]
@@ -90,10 +66,10 @@ RIj2P/wtcJRnuztcszBJsnt7NQ==
         [HttpGet]
         [Route("private-key-jwt")]
         [ServiceFilter(typeof(LogActionEntryAttribute))]
-        public IActionResult PrivateKeyJwt()
+        public async Task<IActionResult> PrivateKeyJwt()
         {
             var model = new PrivateKeyJwtModel();
-            SetDefaults(model);
+            await SetDefaults(model);
             return View(model);
         }
 
@@ -102,36 +78,24 @@ RIj2P/wtcJRnuztcszBJsnt7NQ==
         [ServiceFilter(typeof(LogActionEntryAttribute))]
         public IActionResult PrivateKeyJwt(PrivateKeyJwtModel model)
         {
-            var privateKeyFormatted = FormatPrivateKey(model.PrivateKey);
+            var privateKeyFormatted = model.PrivateKey.FormatPrivateKey();
             var privateKeyJwt = new PrivateKeyJwt(privateKeyFormatted);
-            model.ClientAssertion = privateKeyJwt.Generate(model.Issuer, model.Audience, model.Jti, model.ExpiryMinutes);
+            model.ClientAssertion = privateKeyJwt.Generate(model.Issuer, model.Audience, model.Jti, model.ExpiryMinutes, model.Kid);
             model.ClientAssertionClaims = model.ClientAssertion.GetTokenClaims();
 
             return View(model);
         }
 
-        /// <summary>
-        /// Apply formatting to the provided private key.
-        /// </summary>
-        /// <param name="privateKey">Raw private key</param>
-        /// <returns></returns>
-        private static string FormatPrivateKey(string privateKey)
-        {
-            return privateKey
-                .Replace("-----BEGIN PRIVATE KEY-----", "")
-                .Replace("-----END PRIVATE KEY-----", "")
-                .Replace("\r\n", "")
-                .Trim();
-        }
-
-        private void SetDefaults(PrivateKeyJwtModel model)
+        private async Task SetDefaults(PrivateKeyJwtModel model)
         {
             var sp = _config.GetSoftwareProductConfig();
             var reg = _config.GetRegisterConfig();
+            var tokenEndpoint = await _cacheManager.GetRegisterTokenEndpoint(reg.OidcDiscoveryUri);
 
             model.Issuer = sp.SoftwareProductId;
-            model.Audience = reg.TokenEndpoint;
-            model.PrivateKey = DEFAULT_PRIVATE_KEY;
+            model.Audience = tokenEndpoint;
+            model.PrivateKey = Constants.DEFAULT_PRIVATE_KEY;
+            model.Kid = Constants.DEFAULT_KEY_ID;
             model.ExpiryMinutes = 10;
         }
     }