+++ Source/CDR.DataRecipient.Web/Extensions/ExceptionExtensions.cs
@@ -1,0 +1,25 @@
+﻿using System;
+using System.Runtime.Serialization;
+
+namespace CDR.DataRecipient.Web.Extensions
+{
+    [Serializable]
+    public class CustomException : Exception
+    {
+        public CustomException()
+        {
+        }
+
+        public CustomException(string message) : base(message)
+        {
+        }
+
+        public CustomException(string message, Exception innerException) : base(message, innerException)
+        {
+        }
+
+        protected CustomException(SerializationInfo info, StreamingContext context) : base(info, context)
+        {
+        }
+    }
+}