+++ Source/CDR.DataRecipient/Repository/IDataHoldersRepository.cs
@@ -8,8 +8,9 @@ namespace CDR.DataRecipient.Repository
     {
         Task<IEnumerable<DataHolderBrand>> GetDataHolderBrands();
         Task<DataHolderBrand> GetDataHolderBrand(string brandId);
+        Task<DataHolderBrand> GetDHBrandById(string brandId);
         Task DataHolderBrandsDelete();
-        Task<(int, int)> AggregateDataHolderBrands(IEnumerable<DataHolderBrand> dataHolderBrands);
+        Task<(int, int)> AggregateDataHolderBrands(IList<DataHolderBrand> dataHolderBrands);
         Task PersistDataHolderBrands(IEnumerable<DataHolderBrand> dataHolderBrands);
     }
 }