+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.0.0] - 2022-08-02
+### Added
+- Azure functions to perform Data Holder discovery by polling the Get Data Holder Brands API of the Register.
+
+### Changed
+- First version of the Mock Data Recipient deployed into the CDR Sandbox.
+
 ## [0.2.1] - 2022-06-09
 ### Changed
 - Build and Test action to archive test results. End to end tests now included in test report.